+++ docs/pages/release_notes.md
@@ -14,7 +14,16 @@ This is a {{ site.pmd.release_type }} release.
 
 ### New and noteworthy
 
+#### Modified rules
+
+* The Java rule {% rule java/bestpractices/UnusedPrivateField %} has a new property `annotations`.
+  This is a list of fully qualified names of the annotation types that should be reported anyway. If an unused field
+  has any of these annotations, then it is reported. If it has any other annotation, then it is still considered 
+  to be used and is not reported.
+
 ### Fixed Issues
+* java-bestpractices
+    * [#4166](https://github.com/pmd/pmd/issues/4166): \[java] UnusedPrivateField doesn't find annotated unused private fields anymore
 
 ### API Changes